% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputed.R
\name{express_precomputed_tsne}
\alias{express_precomputed_tsne}
\title{Express Precomputed t-SNE Plot}
\usage{
express_precomputed_tsne(
  dataset,
  interactive = FALSE,
  just_return_data = FALSE
)
}
\arguments{
\item{dataset}{A string specifying the name of the dataset to be retrieved. The dataset must be one of the available datasets listed by \code{express_available_datasets()}.}

\item{interactive}{A logical value indicating whether the plot should be interactive. Currently, this parameter is not used in the function.}

\item{just_return_data}{A logical value indicating whether to return only the dataset without generating a plot. Defaults to FALSE.}
}
\value{
If \code{just_return_data} is TRUE, a data.table containing the dataset is returned. If \code{just_return_data} is FALSE and the dataset type is "tsne", a ggplot2 object representing the t-SNE plot is returned. If the dataset type is not "tsne", an error is raised.
}
\description{
This function retrieves and visualizes precomputed t-SNE data from a specified dataset.
It fetches the data from a URL associated with the dataset name provided, and then either
returns the data directly or creates a ggplot2 visualization of the t-SNE dimensions.
}
\examples{
# Retrieve and plot the t-SNE data
express_precomputed_tsne("example_dataset")

# Retrieve only the t-SNE data without plotting
data <- express_precomputed_tsne("example_dataset", just_return_data = TRUE)
}
