% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/express_tsne.R
\name{express_sample_relatedness}
\alias{express_sample_relatedness}
\title{Title}
\usage{
express_sample_relatedness(
  data,
  method = "TSNE",
  gene_selector = "most_variable",
  sample_metadata = NULL,
  colour_by = NULL,
  perplexity = 30
)
}
\arguments{
\item{perplexity}{if NULL, autocalculate}
}
\description{
Title
}
